# specs/001-ai-image-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI 图像生成平台 API
  version: 1.0.0
  description: |
    v1 公开 REST 契约。AI 路由由 gateway-service 实现，其余由 platform-api 实现。
servers:
  - url: https://api.example.com
    description: 生产
  - url: http://localhost:8080
    description: platform-api
  - url: http://localhost:8081
    description: gateway-service

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code: { type: integer, example: 200 }
        message: { type: string }
        data: {}
        timestamp: { type: integer, format: int64 }
    LoginRequest:
      type: object
      required: [account, password]
      properties:
        account: { type: string, description: 手机号或邮箱 }
        password: { type: string, minLength: 8, maxLength: 20 }
        rememberMe: { type: boolean, default: false }
    RegisterRequest:
      type: object
      required: [password, nickname, verifyCode]
      properties:
        phone: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        nickname: { type: string }
        verifyCode: { type: string }
        inviteCode: { type: string, maxLength: 10 }
    TokenData:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
    ImageGenerateRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string, minLength: 1 }
        size: { type: string, enum: ['1024x1024', '1792x1024', '1024x1792'], default: '1024x1024' }
        n: { type: integer, minimum: 1, maximum: 1, default: 1 }
        sessionId: { type: string, format: uuid }
    ImageGenerateResponse:
      type: object
      properties:
        taskId: { type: string }
        imageUrl: { type: string, format: uri }
        remainingQuota: { type: integer }
    ChatRequest:
      type: object
      required: [messages, sessionId]
      properties:
        sessionId: { type: string, format: uuid }
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role: { type: string, enum: [user, assistant, system] }
              content: { type: string }
    CreateOrderResponse:
      type: object
      properties:
        orderNo: { type: string }
        codeUrl: { type: string, description: 微信 Native 二维码链接 }
        expiresAt: { type: string, format: date-time }

paths:
  /api/v1/auth/send-sms:
    post:
      tags: [Auth]
      summary: 发送手机验证码
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, pattern: '^1[3-9]\d{9}$' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }

  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: 用户注册
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '200': { description: OK }

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 登录
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenData' }

  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新 Token
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /api/v1/ai/chat:
    post:
      tags: [AI-Gateway]
      summary: 文字对话（SSE）
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatRequest' }
      responses:
        '200':
          description: text/event-stream
          content:
            text/event-stream:
              schema: { type: string }

  /api/v1/ai/image/generate:
    post:
      tags: [AI-Gateway]
      summary: 生成图片（先扣后调）
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImageGenerateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ImageGenerateResponse' }
        '402': { description: 次数不足 }
        '409': { description: 已有进行中的生图任务 }

  /api/v1/ai/sessions:
    get:
      tags: [AI]
      summary: 会话列表
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /api/v1/ai/sessions/{id}/messages:
    get:
      tags: [AI]
      summary: 会话消息（分页）
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: OK }

  /api/v1/user/quota:
    get:
      tags: [User]
      summary: 剩余次数
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /api/v1/user/images:
    get:
      tags: [User]
      summary: 个人图库
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        '200': { description: OK }

  /api/v1/user/invite:
    get:
      tags: [User]
      summary: 邀请码与链接
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /api/v1/templates:
    get:
      tags: [Template]
      summary: 模版广场（仅 APPROVED；游客可访问）
      parameters:
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: keyword
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [hot, latest] }
      responses:
        '200': { description: OK }

  /api/v1/templates/publish:
    post:
      tags: [Template]
      summary: 发布模版
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /api/v1/templates/{id}/use:
    post:
      tags: [Template]
      summary: 使用模版
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }

  /api/v1/pay/packages:
    get:
      tags: [Pay]
      summary: 充值套餐列表
      responses:
        '200': { description: OK }

  /api/v1/pay/create-order:
    post:
      tags: [Pay]
      summary: 创建 Native 支付订单
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [packageId]
              properties:
                packageId: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CreateOrderResponse' }

  /api/v1/pay/wx-notify:
    post:
      tags: [Pay]
      summary: 微信支付回调（无 JWT）
      responses:
        '200': { description: 微信要求的成功应答 }

  /api/v1/pay/order/{orderNo}:
    get:
      tags: [Pay]
      summary: 查询订单
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderNo
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  /api/v1/admin/templates/pending:
    get:
      tags: [Admin]
      summary: 待审核模版
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /api/v1/admin/templates/{id}/audit:
    post:
      tags: [Admin]
      summary: 审核模版
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [APPROVE, REJECT] }
                reason: { type: string }
      responses:
        '200': { description: OK }

  /internal/quota/reserve:
    post:
      tags: [Internal]
      summary: 预扣 1 次（gateway 调用）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: integer }
                reservationId: { type: string }
      responses:
        '200': { description: OK }
        '402': { description: 余额不足 }
        '409': { description: in-flight }

  /internal/quota/commit:
    post:
      tags: [Internal]
      summary: 确认扣减
      responses:
        '200': { description: OK }

  /internal/quota/rollback:
    post:
      tags: [Internal]
      summary: 回滚预扣
      responses:
        '200': { description: OK }